Refactor each toolbox section in to separate modules
